backend/requirements.txt
fastapi
uvicorn
requests
python-jose
pymongo
python-dotenv
PyJWT
cryptography